bosss

+++ app/src/main/java/sega/fastnetwork/test/model/Location.kt
@@ -4,24 +4,23 @@ import android.os.Parcel
 import android.os.Parcelable
 
 /**
- * Created by sega4 on 15/08/2017.
+ * Created by cc on 9/13/2017.
  */
-
 class Location() : Parcelable {
     var point: String? = null
     var address: String? = null
-   /* var coordinates : IntArray? = null*/
-   var coordinates : ArrayList<Int>? = ArrayList()
+    /* var coordinates : IntArray? = null*/
+    var coordinates : ArrayList<Double>? = ArrayList()
     constructor(parcel: Parcel) : this() {
         point = parcel.readString()
         address = parcel.readString()
-      /*  coordinates = parcel.createIntArray()*/
+        /*  coordinates = parcel.createIntArray()*/
     }
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         parcel.writeString(point)
         parcel.writeString(address)
-      /*  parcel.writeIntArray(coordinates)*/
+        /*  parcel.writeIntArray(coordinates)*/
     }
 
     override fun describeContents(): Int {
